chore: crate the https outcalls adapter only by passing a config (#1076)
