+++ rs/https_outcalls/adapter/src/lib.rs
@@ -13,17 +13,9 @@ mod metrics;
 
 pub use cli::Cli;
 pub use config::{Config, IncomingSource};
-pub use rpc_server::{CanisterHttp, CanisterRequestBody};
+pub use rpc_server::CanisterHttp;
 
 use futures::{Future, Stream};
-use http_body_util::Full;
-use hyper::body::Bytes;
-use hyper_rustls::HttpsConnectorBuilder;
-use hyper_socks2::SocksConnector;
-use hyper_util::{
-    client::legacy::{connect::HttpConnector, Client},
-    rt::TokioExecutor,
-};
 use ic_https_outcalls_service::canister_http_service_server::CanisterHttpServiceServer;
 use ic_logger::ReplicaLogger;
 use ic_metrics::MetricsRegistry;
@@ -39,44 +31,7 @@ pub struct AdapterServer(Router<Identity>);
 
 impl AdapterServer {
     pub fn new(config: Config, logger: ReplicaLogger, metrics: &MetricsRegistry) -> Self {
-        // Socks client setup
-        let mut http_connector = HttpConnector::new();
-        http_connector.enforce_http(false);
-        http_connector
-            .set_connect_timeout(Some(Duration::from_secs(config.http_connect_timeout_secs)));
-        // The proxy connnector requires a the URL scheme to be specified. I.e socks5://
-        // Config validity check ensures that url includes scheme, host and port.
-        // Therefore the parse 'Uri' will be in the correct format. I.e socks5://somehost.com:1080
-        let proxy_connector = SocksConnector {
-            proxy_addr: config
-                .socks_proxy
-                .parse()
-                .expect("Failed to parse socks url."),
-            auth: None,
-            connector: http_connector.clone(),
-        };
-        let https_connector = HttpsConnectorBuilder::new()
-            .with_native_roots()
-            .expect("Failed to set native roots")
-            .https_only()
-            .enable_http1()
-            .wrap_connector(proxy_connector);
-        let socks_client =
-            Client::builder(TokioExecutor::new()).build::<_, Full<Bytes>>(https_connector);
-
-        // Https client setup.
-        let builder = HttpsConnectorBuilder::new()
-            .with_native_roots()
-            .expect("Failed to set native roots");
-        #[cfg(not(feature = "http"))]
-        let builder = builder.https_only();
-        #[cfg(feature = "http")]
-        let builder = builder.https_or_http();
-
-        let builder = builder.enable_http1();
-        let https_client = Client::builder(TokioExecutor::new())
-            .build::<_, Full<Bytes>>(builder.wrap_connector(http_connector));
-        let canister_http = CanisterHttp::new(https_client, socks_client, logger, metrics);
+        let canister_http = CanisterHttp::new(config.clone(), logger, metrics);
 
         Self(
             Server::builder()
