+++ rs/tests/driver/BUILD.bazel
@@ -127,7 +127,6 @@ rust_library(
         "@crate_index//:http_0_2_12",
         "@crate_index//:humantime",
         "@crate_index//:humantime-serde",
-        "@crate_index//:hyper-rustls",
         "@crate_index//:hyper_0_14_27",
         "@crate_index//:ic-agent",
         "@crate_index//:ic-btc-interface",
